refactor(model): split components and architectures
