+++ src/tab_foundry/model/__init__.py
@@ -1,5 +1,5 @@
 """Model modules."""
 
-from .tabiclv2 import TabICLv2Classifier, TabICLv2Regressor
+from .architectures import TabICLv2Classifier, TabICLv2Regressor
 
 __all__ = ["TabICLv2Classifier", "TabICLv2Regressor"]
