+++ src/tab_foundry/model/architectures/__init__.py
@@ -1,0 +1,5 @@
+"""Architecture family implementations."""
+
+from .tabiclv2 import TabICLv2Classifier, TabICLv2Regressor
+
+__all__ = ["TabICLv2Classifier", "TabICLv2Regressor"]
