+++ src/tab_foundry/model/components/__init__.py
@@ -1,0 +1,3 @@
+"""Reusable model components."""
+
+__all__: list[str] = []
