+++ src/tab_foundry/model/components/many_class.py
