+++ tests/benchmark/test_checkpoint_classifier.py
@@ -9,7 +9,7 @@ import torch
 from torch import nn
 
 import tab_foundry.bench.checkpoint as checkpoint_classifier
-from tab_foundry.model.tabiclv2 import ClassificationOutput
+from tab_foundry.model.architectures.tabiclv2 import ClassificationOutput
 from tab_foundry.types import TaskBatch
